requirement.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
Flask-MQTT==1.1.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
python-dateutil==2.8.2
python-dotenv==0.14.0
pytz==2022.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.1.2
zipp==3.8.0